+++ R/SwimmeR-deprecated.R
@@ -3,4 +3,5 @@
 #' These functions still work but will be removed (defunct) in upcoming versions.
 #'
 #' @name SwimmeR-deprecated
+#'
 NULL